- Move `Attributes` and `Messages` to Add `NsbBridgePlayground.Common` - Update references
