+++ NsbBridgePlayground.Common/Messages/Commands/CreateOrder.cs
@@ -1,0 +1,9 @@
+﻿using NsbBridgePlayground.Common.Attributes;
+
+namespace NsbBridgePlayground.Common.Messages.Commands;
+
+[NsbCommand(Endpoints.OrderProcessor)]
+public class CreateOrder
+{
+  public Guid Id { get; set; }
+}
