+++ NsbBridgePlayground.Notifier/Handlers/OrderCreatedHandler.cs
@@ -1,5 +1,5 @@
 using Microsoft.Extensions.Logging;
-using NsbBridgePlayground.Bootstrap.Messages.Events;
+using NsbBridgePlayground.Common.Messages.Events;
 using NServiceBus;
 
 namespace NsbBridgePlayground.Notifier.Handlers;
