+++ NsbBridgePlayground.OrderProcessor/Handlers/CreateOrderHandler.cs
@@ -1,6 +1,6 @@
 using Microsoft.Extensions.Logging;
-using NsbBridgePlayground.Bootstrap.Messages.Commands;
-using NsbBridgePlayground.Bootstrap.Messages.Events;
+using NsbBridgePlayground.Common.Messages.Commands;
+using NsbBridgePlayground.Common.Messages.Events;
 using NServiceBus;
 
 namespace NsbBridgePlayground.OrderProcessor.Handlers;
