+++ NsbBridgePlayground.Sender/Program.cs
@@ -2,8 +2,8 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using NsbBridgePlayground.Bootstrap.Infrastructure;
-using NsbBridgePlayground.Bootstrap.Messages.Commands;
 using NsbBridgePlayground.Common;
+using NsbBridgePlayground.Common.Messages.Commands;
 using NServiceBus;
 
 namespace NsbBridgePlayground.Sender;
